refactor(lib): extract restore_from_repository
