+++ lib/restore_bundled_with.rb
@@ -1,6 +1,7 @@
 require 'git'
 require 'logger'
 
+require 'restore_bundled_with/restore_from_repository'
 require 'restore_bundled_with/error'
 require 'restore_bundled_with/lock'
 require 'restore_bundled_with/repository'
