+++ lib/restore_bundled_with/restore_from_repository.rb
@@ -1,0 +1,6 @@
+require 'restore_bundled_with/restore_from_repository/error'
+require 'restore_bundled_with/restore_from_repository/file'
+require 'restore_bundled_with/restore_from_repository/repository'
+
+module RestoreFromRepository
+end
