+++ lib/restore_bundled_with/restore_from_repository/error.rb
@@ -1,0 +1,3 @@
+module RestoreFromRepository
+  class NoInputError < StandardError; end
+end
